move apply.stateproof tests
