+++ ledger/internal/eval_test.go
@@ -31,14 +31,11 @@ import (
 	"github.com/algorand/go-algorand/config"
 	"github.com/algorand/go-algorand/crypto"
 	"github.com/algorand/go-algorand/crypto/merklesignature"
-	"github.com/algorand/go-algorand/crypto/stateproof"
 	"github.com/algorand/go-algorand/data/basics"
 	"github.com/algorand/go-algorand/data/bookkeeping"
-	"github.com/algorand/go-algorand/data/stateproofmsg"
 	"github.com/algorand/go-algorand/data/transactions"
 	"github.com/algorand/go-algorand/data/transactions/logic"
 	"github.com/algorand/go-algorand/data/transactions/verify"
-	"github.com/algorand/go-algorand/ledger/apply"
 	"github.com/algorand/go-algorand/ledger/ledgercore"
 	ledgertesting "github.com/algorand/go-algorand/ledger/testing"
 	"github.com/algorand/go-algorand/protocol"
@@ -198,204 +195,6 @@ func TestEvalAppAllocStateWithTxnGroup(t *testing.T) {
 	state := ad.AppParams[1].GlobalState
 	require.Equal(t, basics.TealValue{Type: basics.TealBytesType, Bytes: string(addr[:])}, state["caller"])
 	require.Equal(t, basics.TealValue{Type: basics.TealBytesType, Bytes: string(addr[:])}, state["creator"])
-}
-
-func TestCowStateProofV34(t *testing.T) {
-	partitiontest.PartitionTest(t)
-	a := require.New(t)
-
-	var spType protocol.StateProofType
-	var stateProof stateproof.StateProof
-	var atRound basics.Round
-	var validate bool
-	msg := stateproofmsg.Message{}
-
-	const version = protocol.ConsensusV34
-
-	accts0 := ledgertesting.RandomAccounts(20, true)
-	blocks := make(map[basics.Round]bookkeeping.BlockHeader)
-	blockErr := make(map[basics.Round]error)
-	ml := mockLedger{balanceMap: accts0, blocks: blocks, blockErr: blockErr}
-	c0 := makeRoundCowState(
-		&ml, bookkeeping.BlockHeader{}, config.Consensus[protocol.ConsensusV34],
-		0, ledgercore.AccountTotals{}, 0)
-
-	spType = protocol.StateProofType(1234) // bad stateproof type
-	stateProofTx := transactions.StateProofTxnFields{
-		StateProofType: spType,
-		StateProof:     stateProof,
-		Message:        msg,
-	}
-	err := apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.ErrorIs(err, apply.ErrStateProofTypeNotSupported)
-
-	stateProofTx.StateProofType = protocol.StateProofBasic
-	// stateproof txn doesn't confirm the next state proof round. expected is in the past
-	validate = true
-	stateProofTx.Message.LastAttestedRound = uint64(16)
-	c0.SetStateProofNextRound(8)
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.ErrorIs(err, apply.ErrExpectedDifferentStateProofRound)
-	c0.SetStateProofNextRound(32)
-
-	// stateproof txn doesn't confirm the next state proof round. expected is in the future
-	validate = true
-	stateProofTx.Message.LastAttestedRound = uint64(16)
-	c0.SetStateProofNextRound(32)
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.ErrorIs(err, apply.ErrExpectedDifferentStateProofRound)
-
-	// no atRound and lastAttested block
-	stateProofTx.Message.LastAttestedRound = 32
-	noBlockErr := errors.New("no block")
-	blockErr[atRound] = noBlockErr
-	blockErr[basics.Round(stateProofTx.Message.LastAttestedRound)] = noBlockErr
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.ErrorIs(err, noBlockErr)
-	delete(blockErr, atRound)
-
-	atRoundBlock := bookkeeping.BlockHeader{}
-	atRoundBlock.CurrentProtocol = version
-	blocks[atRound] = atRoundBlock
-
-	// no spRnd block
-	noBlockErr = errors.New("no block")
-	blockErr[32] = noBlockErr
-	stateProofTx.Message.LastAttestedRound = 32
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.ErrorIs(err, noBlockErr)
-
-	// no votersRnd block
-	// this is slightly a mess of things that don't quite line up with likely usage
-	validate = true
-	var spHdr bookkeeping.BlockHeader
-	spHdr.CurrentProtocol = "TestCowStateProof"
-	spHdr.Round = 1
-	proto := config.Consensus[spHdr.CurrentProtocol]
-	proto.StateProofInterval = 2
-	config.Consensus[spHdr.CurrentProtocol] = proto
-	blocks[spHdr.Round] = spHdr
-
-	spHdr.Round = 15
-	blocks[spHdr.Round] = spHdr
-	stateProofTx.Message.LastAttestedRound = uint64(spHdr.Round)
-	c0.SetStateProofNextRound(15)
-	blockErr[13] = noBlockErr
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.Contains(err.Error(), "no block")
-	delete(blockErr, 13)
-
-	// check the happy flow - we should fail only on crypto
-	atRound = 800
-	spHdr = bookkeeping.BlockHeader{}
-	spHdr.CurrentProtocol = version
-	blocks[basics.Round(2*config.Consensus[version].StateProofInterval)] = spHdr
-
-	votersHdr := bookkeeping.BlockHeader{}
-	votersHdr.CurrentProtocol = version
-	stateproofTracking := bookkeeping.StateProofTrackingData{
-		StateProofVotersCommitment:  []byte{0x1}[:],
-		StateProofOnlineTotalWeight: basics.MicroAlgos{Raw: 5},
-	}
-	votersHdr.StateProofTracking = make(map[protocol.StateProofType]bookkeeping.StateProofTrackingData)
-	votersHdr.StateProofTracking[protocol.StateProofBasic] = stateproofTracking
-
-	blocks[basics.Round(config.Consensus[version].StateProofInterval)] = votersHdr
-	atRoundBlock = bookkeeping.BlockHeader{}
-	atRoundBlock.CurrentProtocol = version
-	blocks[atRound] = atRoundBlock
-
-	stateProofTx.Message.LastAttestedRound = 2 * config.Consensus[version].StateProofInterval
-	stateProofTx.StateProof.SignedWeight = 100
-	c0.SetStateProofNextRound(basics.Round(2 * config.Consensus[version].StateProofInterval))
-
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.Contains(err.Error(), "crypto error")
-}
-
-func TestCowStateProof(t *testing.T) {
-	partitiontest.PartitionTest(t)
-	a := require.New(t)
-
-	var spType protocol.StateProofType
-	var stateProof stateproof.StateProof
-	atRound := basics.Round(600)
-	var validate bool
-	msg := stateproofmsg.Message{}
-
-	accts0 := ledgertesting.RandomAccounts(20, true)
-	blocks := make(map[basics.Round]bookkeeping.BlockHeader)
-	blockErr := make(map[basics.Round]error)
-	ml := mockLedger{balanceMap: accts0, blocks: blocks, blockErr: blockErr}
-	c0 := makeRoundCowState(
-		&ml, bookkeeping.BlockHeader{}, config.Consensus[protocol.ConsensusFuture],
-		0, ledgercore.AccountTotals{}, 0)
-	ml.stateProofVerification = make(map[basics.Round]*ledgercore.StateProofVerificationContext)
-
-	spType = protocol.StateProofType(1234) // bad stateproof type
-	stateProofTx := transactions.StateProofTxnFields{
-		StateProofType: spType,
-		StateProof:     stateProof,
-		Message:        msg,
-	}
-	err := apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.ErrorIs(err, apply.ErrStateProofTypeNotSupported)
-
-	stateProofTx.StateProofType = protocol.StateProofBasic
-	// stateproof txn doesn't confirm the next state proof round. expected is in the past
-	validate = true
-	stateProofTx.Message.LastAttestedRound = uint64(16)
-	c0.SetStateProofNextRound(8)
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.ErrorIs(err, apply.ErrExpectedDifferentStateProofRound)
-	c0.SetStateProofNextRound(32)
-
-	// stateproof txn doesn't confirm the next state proof round. expected is in the future
-	validate = true
-	stateProofTx.Message.LastAttestedRound = uint64(16)
-	c0.SetStateProofNextRound(32)
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.ErrorIs(err, apply.ErrExpectedDifferentStateProofRound)
-
-	atRoundBlock := bookkeeping.BlockHeader{}
-	atRoundBlock.CurrentProtocol = protocol.ConsensusFuture
-	blocks[atRound] = atRoundBlock
-
-	validate = true
-	// no Verification Context for rounds 32
-	stateProofTx.Message.LastAttestedRound = 32
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.ErrorIs(err, ErrVerificationContextNotFound)
-	delete(blockErr, atRound)
-
-	// the behavior has changed and we no longer require the voters blockheader to verify the transaction
-	// still, this test should assure the error returned is the one expected and not "no block"
-	noBlockErr := errors.New("no block")
-
-	// removing blocks for the ledger so if apply.stateproof uses the tracker it should pass
-	c0.SetStateProofNextRound(512)
-	blockErr[512] = noBlockErr
-	blockErr[256] = noBlockErr
-	stateProofTx.Message.LastAttestedRound = 512
-	stateProofTx.StateProof.SignedWeight = 100
-	ml.stateProofVerification[basics.Round(stateProofTx.Message.LastAttestedRound)] = &ledgercore.StateProofVerificationContext{
-		LastAttestedRound: basics.Round(stateProofTx.Message.LastAttestedRound),
-		VotersCommitment:  []byte{0x1}[:],
-		OnlineTotalWeight: basics.MicroAlgos{Raw: 5},
-		Version:           protocol.ConsensusFuture,
-	}
-
-	// crypto verification should fail since it is not a valid stateproof
-	err = apply.StateProof(stateProofTx, atRound, c0, validate)
-	a.Error(err)
-	a.Contains(err.Error(), "crypto error")
-
-	a.Equal(basics.Round(512), c0.mods.StateProofNext)
-	// transaction should be applied without stateproof validation (no context, blockheader or valid stateproof needed as it represents a node catching up)
-	err = apply.StateProof(stateProofTx, atRound, c0, false)
-	a.NoError(err)
-	// make sure that the StateProofNext was updated correctly after applying
-	a.Equal(basics.Round(512+config.Consensus[protocol.ConsensusFuture].StateProofInterval), c0.mods.StateProofNext)
 }
 
 // a couple trivial tests that don't need setup
